ch.02: Let us not be too assuming about the gender of "each miner"
It is too assuming to suppose that the gender of "each miner" is masculine. Hence I propose to use [singular they](https://en.wikipedia.org/wiki/Singular_they). More on singular they [here](https://www.grammarly.com/blog/use-the-singular-they/).
'He' was substituted with 'they' only when it refers to "each miner", i.e. miners in general, and not when it refers to Jing. In the latter case I have left he unchanged.